vpn-service/bot/requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
qrcode[pil]==7.4.2
Pillow==10.1.0
structlog==23.2.0
httpx==0.25.2
asyncio-mqtt==0.16.1
python-dateutil==2.8.2

# Зависимости для PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
